issue-7987-gwt-2.12.2: reverting the addition of a RuntimeException custom field serializer;
It breaks a number of tests, specifically in RPCSuite. The reason given
in the serializer's comment was that there was a lack of symmetry for
the NullPointerException which is treated special by GWT RPC, namely as
a JsException.

But should RuntimeException serializability be required by an
application, the application could still add such a CSF on its own.